requirement.txt
streamlit
PyPDF2
langchain
langchain-community
langchain-huggingface
langchain-groq
langchain-text-splitters
chromadb
sentence-transformers
python-dotenv